edit variable in memory initial implementation
It looks like it works? it doesnt cause an error other than string overflow as far as i can tell. I tested scheduler with it and it looks like the values got edited naman. I'm not sure about the screen -c because it just shows no output so I havent tested that